modules unit tests

+++ src/modules/modules/module.service.ts
@@ -1,6 +1,11 @@
-import { BadRequestException, Inject, Injectable, NotFoundException } from '@nestjs/common';
-import upload from 'src/config/upload';
-import { LessThan, LessThanOrEqual, MoreThan, MoreThanOrEqual, Repository } from 'typeorm';
+import {
+  BadRequestException,
+  Inject,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
+import upload from '../../config/upload';
+import { MoreThan, Repository } from 'typeorm';
 import StorageProvider from '../storage/storage-provider-model';
 import { CreateModuleDTO } from './dtos/create-module.dto';
 import { FindModulesQuery } from './dtos/find-modules-query.dto';
@@ -16,23 +21,22 @@ export class ModulesService {
     private modulesRepository: Repository<Module>,
 
     @Inject('StorageProvider')
-    private storageProvider: StorageProvider
-  ) { }
+    private storageProvider: StorageProvider,
+  ) {}
 
-  async create({
-    title,
-    order,
-    icon
-  }: CreateModuleDTO): Promise<Module> {
-    const moduleInposition = await this.modulesRepository.findOne({ where: { order } });
+  async create({ title, order, icon }: CreateModuleDTO): Promise<Module> {
+    const moduleInposition = await this.modulesRepository.findOne({
+      where: { order },
+    });
 
     if (moduleInposition) {
-      if (upload.driver === 'disk')
-        await this.storageProvider.deleteFile(icon)
-      throw new BadRequestException('Já existe um módulo nessa posição. Você precisa reordenar os módulos.')
+      if (upload.driver === 'disk') await this.storageProvider.deleteFile(icon);
+      throw new BadRequestException(
+        'Já existe um módulo nessa posição. Você precisa reordenar os módulos.',
+      );
     }
 
-    const imageUrl = await this.storageProvider.saveFile(icon)
+    const imageUrl = await this.storageProvider.saveFile(icon);
 
     const module = this.modulesRepository.create({
       title,
@@ -48,10 +52,13 @@ export class ModulesService {
 
     if (!module) throw new NotFoundException('Módulo não encontrado.');
 
-    await this.modulesRepository.update({ id: module.id }, {
-      ...module,
-      title,
-    })
+    await this.modulesRepository.update(
+      { id: module.id },
+      {
+        ...module,
+        title,
+      },
+    );
 
     return {
       ...module,
@@ -60,64 +67,101 @@ export class ModulesService {
   }
 
   async updateIcon(id: string, icon: string): Promise<Module> {
-    if (icon === '') throw new BadRequestException('Conteúdo enviado não é uma imagem.')
+    if (icon === '')
+      throw new BadRequestException('Conteúdo enviado não é uma imagem.');
 
     const module = await this.modulesRepository.findOne({ where: { id } });
 
     if (!module) {
-      if (upload.driver === 'disk')
-        await this.storageProvider.deleteFile(icon)
+      if (upload.driver === 'disk') await this.storageProvider.deleteFile(icon);
       throw new NotFoundException('Módulo não encontrado.');
     }
 
-    await this.storageProvider.deleteFile(module.icon_url)
+    await this.storageProvider.deleteFile(module.icon_url);
 
-    const imageUrl = await this.storageProvider.saveFile(icon)
+    const imageUrl = await this.storageProvider.saveFile(icon);
 
-    await this.modulesRepository.update({ id }, {
-      icon_url: imageUrl,
-    })
+    await this.modulesRepository.update(
+      { id },
+      {
+        icon_url: imageUrl,
+      },
+    );
 
     return {
       ...module,
-      icon_url: imageUrl
-    }
+      icon_url: imageUrl,
+    };
   }
 
   async reorder({ id, order }: ReorderModulesDTO): Promise<void> {
     const allModules = await this.find({});
 
-    const moduleExists = allModules.modules.filter((module) => module.id === id)[0]
-    const moduleReplaced = allModules.modules.filter((module) => module.order === order)[0]
+    const moduleExists = allModules.modules.filter(
+      (module) => module.id === id,
+    )[0];
+    const moduleReplaced = allModules.modules.filter(
+      (module) => module.order === order,
+    )[0];
 
-    if (!moduleExists) throw new BadRequestException('Um dos módulos especificados não existe. Verifique os dados novamente.')
+    if (!moduleExists)
+      throw new BadRequestException(
+        'Um dos módulos especificados não existe. Verifique os dados novamente.',
+      );
 
-    await this.modulesRepository.update({ id: moduleExists.id }, { ...moduleExists, order: null })
+    await this.modulesRepository.update(
+      { id: moduleExists.id },
+      { ...moduleExists, order: null },
+    );
 
-    await this.modulesRepository.update({ id: moduleReplaced.id }, { ...moduleReplaced, order: null })
+    await this.modulesRepository.update(
+      { id: moduleReplaced.id },
+      { ...moduleReplaced, order: null },
+    );
 
     if (moduleExists.order < order) {
-      const afterModules = allModules.modules.filter((module) => module.order > moduleExists.order && module.order <= order);
+      const afterModules = allModules.modules.filter(
+        (module) => module.order > moduleExists.order && module.order <= order,
+      );
 
       const afterModulesPromises = afterModules.map(async (afterModule) => {
-        await this.modulesRepository.update({ id: afterModule.id }, { ...afterModule, order: afterModule.order - 1 })
-      })
+        await this.modulesRepository.update(
+          { id: afterModule.id },
+          { ...afterModule, order: afterModule.order - 1 },
+        );
+      });
 
-      await this.modulesRepository.update({ id: moduleExists.id }, { ...moduleExists, order: order })
+      await this.modulesRepository.update(
+        { id: moduleExists.id },
+        { ...moduleExists, order: order },
+      );
 
-      await Promise.all(afterModulesPromises)
+      await Promise.all(afterModulesPromises);
     } else {
-      const beforeModules = allModules.modules.filter((module) => module.order < moduleExists.order && module.order > order);
+      const beforeModules = allModules.modules.filter(
+        (module) => module.order < moduleExists.order && module.order > order,
+      );
 
-      const beforeModulesPromises = beforeModules.reverse().map(async (beforeModule) => {
-        await this.modulesRepository.update({ id: beforeModule.id }, { ...beforeModule, order: beforeModule.order + 1 })
-      })
+      const beforeModulesPromises = beforeModules
+        .reverse()
+        .map(async (beforeModule) => {
+          await this.modulesRepository.update(
+            { id: beforeModule.id },
+            { ...beforeModule, order: beforeModule.order + 1 },
+          );
+        });
 
-      await this.modulesRepository.update({ id: moduleExists.id }, { ...moduleExists, order: order })
+      await this.modulesRepository.update(
+        { id: moduleExists.id },
+        { ...moduleExists, order: order },
+      );
 
-      await this.modulesRepository.update({ id: moduleReplaced.id }, { ...moduleReplaced, order: order + 1 })
+      await this.modulesRepository.update(
+        { id: moduleReplaced.id },
+        { ...moduleReplaced, order: order + 1 },
+      );
 
-      await Promise.all(beforeModulesPromises)
+      await Promise.all(beforeModulesPromises);
     }
   }
 
@@ -131,16 +175,19 @@ export class ModulesService {
 
   async find({ count, page }: FindModulesQuery): Promise<ListModuleResponse> {
     if (count && page) {
-      page = page - 1
+      page = page - 1;
 
-      if (page < 0) throw new BadRequestException('Número da página deve ser maior que zero.')
+      if (page < 0)
+        throw new BadRequestException(
+          'Número da página deve ser maior que zero.',
+        );
 
       const skip = page * count;
       const [modules, total] = await this.modulesRepository.findAndCount({
         order: { order: 'ASC' },
         take: count,
-        skip
-      })
+        skip,
+      });
 
       let next = false;
 
@@ -148,18 +195,25 @@ export class ModulesService {
 
       if (page < totalPages - 1) next = true;
 
-      if (modules.length === 0) throw new BadRequestException('Nenhum módulo encontrado para os filtros selecionados.')
+      if (modules.length === 0)
+        throw new BadRequestException(
+          'Nenhum módulo encontrado para os filtros selecionados.',
+        );
 
       return { modules, next };
     }
 
-    const modules = await this.modulesRepository.find({ order: { order: 'ASC' } });
+    const modules = await this.modulesRepository.find({
+      order: { order: 'ASC' },
+    });
 
     return { modules };
   }
 
   async delete(id: string): Promise<void> {
-    const deletedModule = await this.modulesRepository.findOne({ where: { id } });
+    const deletedModule = await this.modulesRepository.findOne({
+      where: { id },
+    });
 
     if (!deletedModule) throw new NotFoundException('Módulo não encontrado.');
 
@@ -167,12 +221,18 @@ export class ModulesService {
 
     await this.modulesRepository.delete(id);
 
-    const allModules = await this.modulesRepository.find({ where: { order: MoreThan(deletedModule.order) }, order: { order: 'ASC' } });
+    const allModules = await this.modulesRepository.find({
+      where: { order: MoreThan(deletedModule.order) },
+      order: { order: 'ASC' },
+    });
 
     const moduleUpdatePromises = allModules.map(async (module, index) => {
-      let position = deletedModule.order + index;
-      await this.modulesRepository.update({ id: module.id }, { ...module, order: position })
-    })
+      const position = deletedModule.order + index;
+      await this.modulesRepository.update(
+        { id: module.id },
+        { ...module, order: position },
+      );
+    });
 
     await Promise.all(moduleUpdatePromises);
   }
